refactor/#158: response 패키지로 이동
