+++ backend/member-server/src/main/java/com/smiletoegether/memberserver/member/service/dto/response/SignInResponse.java
@@ -1,4 +1,4 @@
-package com.smiletoegether.memberserver.member.service.dto;
+package com.smiletoegether.memberserver.member.service.dto.response;
 
 import com.smiletoegether.memberserver.common.dto.CommonCodeResponse;
