+++ backend/member-server/src/main/java/com/smiletoegether/memberserver/member/service/dto/response/SignUpResponse.java
@@ -1,4 +1,4 @@
-package com.smiletoegether.memberserver.member.service.dto;
+package com.smiletoegether.memberserver.member.service.dto.response;
 
 import com.smiletoegether.memberserver.member.domain.Member;
